cc33xx: More compatible fix for printf warnings
PROD-7804

+++ drivers/net/wireless/ti/cc33xx/cmd.c
@@ -132,9 +132,9 @@ static int __wlcore_cmd_send(struct cc33xx *wl, u16 id, void *buf,
 	case CMD_TEST_MODE:
 	case CMD_BM_READ_DEVICE_INFO:
 		cc33xx_debug(DEBUG_CMD, 
-			"Response len %d, allocated buffer len %lu",
+			"Response len %d, allocated buffer len %d",
 			wl->result_length,
-			res_len);
+			(int)res_len);
 
 		if (!res_len)
 			break; // Response should be discarded  
@@ -168,7 +168,7 @@ static int wlcore_cmd_send_failsafe(struct cc33xx *wl, u16 id, void *buf,
 	int ret = __wlcore_cmd_send(wl, id, buf, len, res_len, true);
 	
 
-	cc33xx_debug(DEBUG_TESTMODE, "CMD# %d, len=%lu", id, len);
+	cc33xx_debug(DEBUG_TESTMODE, "CMD# %d, len=%d", id, (int)len);
 
 	if (ret < 0)
 		goto fail;
